Made newEndCrit and extensive_Init the default behaviour
We reduced the code a bit by deleting alternative types of simulations of the MAS. Now the MAS simulations always behaves as if we had set newEndCrit=true and extensive_Init=true. These were the settings producing the best results (as oppes to setting them to false) and these aspects are not going to be investigated in the future anymore.
The files in the ./Tutorial folder were just updated in names to follow the new naming convention in absence of the two parameters removed.